+++ cpp/open3d/geometry/TriangleMesh.cpp
@@ -70,6 +70,9 @@ TriangleMesh &TriangleMesh::Rotate(const Eigen::Matrix3d &R,
 
 TriangleMesh &TriangleMesh::operator+=(const TriangleMesh &mesh) {
     if (mesh.IsEmpty()) return (*this);
+    bool add_textures = HasTriangleUvs() && HasTextures() &&
+                        HasTriangleMaterialIds() && mesh.HasTriangleUvs() &&
+                        mesh.HasTextures() && mesh.HasTriangleMaterialIds();
     size_t old_vert_num = vertices_.size();
     MeshBase::operator+=(mesh);
     size_t old_tri_num = triangles_.size();
@@ -92,11 +95,30 @@ TriangleMesh &TriangleMesh::operator+=(const TriangleMesh &mesh) {
     if (HasAdjacencyList()) {
         ComputeAdjacencyList();
     }
-    if (HasTriangleUvs() || HasTextures() || HasTriangleMaterialIds()) {
-        utility::LogError(
-                "[TriangleMesh] copy of uvs and texture and per-triangle "
-                "material ids is not implemented "
-                "yet");
+    if (add_textures) {
+        size_t old_tri_uv_num = triangle_uvs_.size();
+        triangle_uvs_.resize(old_tri_uv_num + mesh.triangle_uvs_.size());
+        for (size_t i = 0; i < mesh.triangle_uvs_.size(); i++) {
+            triangle_uvs_[old_tri_uv_num + i] = mesh.triangle_uvs_[i];
+        }
+
+        size_t old_tex_num = textures_.size();
+        textures_.resize(old_tex_num + mesh.textures_.size());
+        for (size_t i = 0; i < mesh.textures_.size(); i++) {
+            textures_[old_tex_num + i] = mesh.textures_[i];
+        }
+
+        size_t old_mat_id_num = triangle_material_ids_.size();
+        triangle_material_ids_.resize(old_mat_id_num +
+                                      mesh.triangle_material_ids_.size());
+        for (size_t i = 0; i < mesh.triangle_material_ids_.size(); i++) {
+            triangle_material_ids_[old_mat_id_num + i] =
+                    mesh.triangle_material_ids_[i] + (int)old_tex_num;
+        }
+    } else {
+        triangle_uvs_.clear();
+        textures_.clear();
+        triangle_material_ids_.clear();
     }
     return (*this);
 }