+++ cpp/open3d/t/pipelines/registration/TransformationEstimation.cpp
@@ -26,7 +26,7 @@
 
 #include "open3d/t/pipelines/registration/TransformationEstimation.h"
 
-#include "open3d/t/pipelines/kernel/ComputePosePointToPlane.h"
+#include "open3d/t/pipelines/kernel/ComputeTransform.h"
 #include "open3d/t/pipelines/kernel/TransformationConverter.h"
 
 namespace open3d {
@@ -37,7 +37,7 @@ namespace registration {
 double TransformationEstimationPointToPoint::ComputeRMSE(
         const geometry::PointCloud &source,
         const geometry::PointCloud &target,
-        CorrespondenceSet &corres) const {
+        const CorrespondenceSet &corres) const {
     core::Device device = source.GetDevice();
     core::Dtype dtype = core::Dtype::Float32;
     source.GetPoints().AssertDtype(dtype);
@@ -50,19 +50,22 @@ double TransformationEstimationPointToPoint::ComputeRMSE(
 
     double error;
     // TODO: Revist to support Float32 and 64 without type conversion.
-    core::Tensor source_select = source.GetPoints().IndexGet({corres.first});
-    core::Tensor target_select = target.GetPoints().IndexGet({corres.second});
+    // TODO: Optimise using kernel.
+    core::Tensor source_select =
+            source.GetPoints().IndexGet({corres.first.Reshape({-1})});
+    core::Tensor target_select =
+            target.GetPoints().IndexGet({corres.second.Reshape({-1})});
 
     core::Tensor error_t = (source_select - target_select);
     error_t.Mul_(error_t);
     error = static_cast<double>(error_t.Sum({0, 1}).Item<float>());
-    return std::sqrt(error / static_cast<double>(corres.second.GetShape()[0]));
+    return std::sqrt(error / static_cast<double>(corres.second.GetLength()));
 }
 
 core::Tensor TransformationEstimationPointToPoint::ComputeTransformation(
         const geometry::PointCloud &source,
         const geometry::PointCloud &target,
-        CorrespondenceSet &corres) const {
+        const CorrespondenceSet &corres) const {
     core::Device device = source.GetDevice();
     core::Dtype dtype = core::Dtype::Float32;
     source.GetPoints().AssertDtype(dtype);
@@ -72,26 +75,10 @@ core::Tensor TransformationEstimationPointToPoint::ComputeTransformation(
                 "Target Pointcloud device {} != Source Pointcloud's device {}.",
                 target.GetDevice().ToString(), device.ToString());
     }
-    core::Tensor source_select = source.GetPoints().IndexGet({corres.first});
-    core::Tensor target_select = target.GetPoints().IndexGet({corres.second});
 
-    // https://ieeexplore.ieee.org/document/88573
-    core::Tensor mux = source_select.Mean({0}, true);
-    core::Tensor muy = target_select.Mean({0}, true);
-    core::Tensor Sxy =
-            ((target_select - muy)
-                     .T()
-                     .Matmul(source_select - mux)
-                     .Div_(static_cast<float>(corres.second.GetShape()[0])));
-    core::Tensor U, D, VT;
-    std::tie(U, D, VT) = Sxy.SVD();
-    core::Tensor S = core::Tensor::Eye(3, dtype, device);
-    if (U.Det() * (VT.T()).Det() < 0) {
-        S[-1][-1] = -1;
-    }
     core::Tensor R, t;
-    R = U.Matmul(S.Matmul(VT));
-    t = muy.Reshape({-1}) - R.Matmul(mux.T()).Reshape({-1});
+    std::tie(R, t) = pipelines::kernel::ComputeRtPointToPoint(
+            source.GetPoints(), target.GetPoints(), corres);
 
     return t::pipelines::kernel::RtToTransformation(R, t);
 }
@@ -99,7 +86,7 @@ core::Tensor TransformationEstimationPointToPoint::ComputeTransformation(
 double TransformationEstimationPointToPlane::ComputeRMSE(
         const geometry::PointCloud &source,
         const geometry::PointCloud &target,
-        CorrespondenceSet &corres) const {
+        const CorrespondenceSet &corres) const {
     core::Device device = source.GetDevice();
     core::Dtype dtype = core::Dtype::Float32;
     source.GetPoints().AssertDtype(dtype);
@@ -111,23 +98,25 @@ double TransformationEstimationPointToPlane::ComputeRMSE(
     }
 
     if (!target.HasPointNormals()) return 0.0;
-
-    core::Tensor source_select = source.GetPoints().IndexGet({corres.first});
-    core::Tensor target_select = target.GetPoints().IndexGet({corres.second});
+    // TODO: Optimise using kernel.
+    core::Tensor source_select =
+            source.GetPoints().IndexGet({corres.first.Reshape({-1})});
+    core::Tensor target_select =
+            target.GetPoints().IndexGet({corres.second.Reshape({-1})});
     core::Tensor target_n_select =
-            target.GetPointNormals().IndexGet({corres.second});
+            target.GetPointNormals().IndexGet({corres.second.Reshape({-1})});
 
     core::Tensor error_t =
             (source_select - target_select).Mul_(target_n_select);
     error_t.Mul_(error_t);
     double error = static_cast<double>(error_t.Sum({0, 1}).Item<float>());
-    return std::sqrt(error / static_cast<double>(corres.second.GetShape()[0]));
+    return std::sqrt(error / static_cast<double>(corres.second.GetLength()));
 }
 
 core::Tensor TransformationEstimationPointToPlane::ComputeTransformation(
         const geometry::PointCloud &source,
         const geometry::PointCloud &target,
-        CorrespondenceSet &corres) const {
+        const CorrespondenceSet &corres) const {
     core::Device device = source.GetDevice();
     core::Dtype dtype = core::Dtype::Float32;
     source.GetPoints().AssertDtype(dtype);
@@ -138,18 +127,10 @@ core::Tensor TransformationEstimationPointToPlane::ComputeTransformation(
                 target.GetDevice().ToString(), device.ToString());
     }
 
-    // Get indexed source and target points and target normals, according to
-    // correspondences.
-    core::Tensor source_indexed =
-            source.GetPoints().IndexGet({corres.first}).To(dtype);
-    core::Tensor target_indexed =
-            target.GetPoints().IndexGet({corres.second}).To(dtype);
-    core::Tensor target_norm_indexed =
-            target.GetPointNormals().IndexGet({corres.second}).To(dtype);
-
     // Get pose {6} from correspondences indexed source and target point cloud.
     core::Tensor pose = pipelines::kernel::ComputePosePointToPlane(
-            source_indexed, target_indexed, target_norm_indexed);
+            source.GetPoints(), target.GetPoints(), target.GetPointNormals(),
+            corres);
 
     // Get transformation {4,4} from pose {6}.
     return pipelines::kernel::PoseToTransformation(pose);