+++ util/ci_utils.sh
@@ -37,7 +37,7 @@ PIP_VER="20.2.4"
 WHEEL_VER="0.35.1"
 STOOLS_VER="50.3.2"
 PYTEST_VER="6.0.1"
-SCIPY_VER="1.4.1"
+SCIPY_VER="1.5.4"
 YAPF_VER="0.30.0"
 
 # Documentation
@@ -519,36 +519,6 @@ build_docs() {
     cd ../docs # To Open3D/docs
     python make_docs.py $DOC_ARGS --pyapi_rst=always --execute_notebooks=never --sphinx --doxygen
     set +x # Echo commands off
-}
-
-install_arm64_dependencies() {
-    apt-get update -q -y
-    apt-get install -y apt-utils build-essential git wget
-    apt-get install -y python3 python3-dev python3-pip python3-virtualenv
-    apt-get install -y xorg-dev libglu1-mesa-dev ccache
-    apt-get install -y libblas-dev liblapack-dev liblapacke-dev libssl-dev
-    apt-get install -y libsdl2-dev libc++-7-dev libc++abi-7-dev libxi-dev
-    apt-get install -y libudev-dev autoconf libtool # librealsense
-    apt-get install -y clang-7
-    update-alternatives --install /usr/bin/clang clang /usr/bin/clang-7 100
-    update-alternatives --install /usr/bin/clang++ clang++ /usr/bin/clang++-7 100
-    update-alternatives --install /usr/bin/cc cc /usr/bin/clang 100
-    update-alternatives --install /usr/bin/c++ c++ /usr/bin/clang++ 100
-    /usr/sbin/update-ccache-symlinks
-    echo 'export PATH="/usr/lib/ccache:$PATH"' | tee -a ~/.bashrc
-    virtualenv --python=$(which python) ${HOME}/venv
-    source ${HOME}/venv/bin/activate
-    which python
-    python --version
-    pip install pytest=="$PYTEST_VER" -U
-    pip install wheel=="$WHEEL_VER" -U
-    # Get pre-compiled CMake
-    wget https://github.com/intel-isl/Open3D/releases/download/v0.11.0/cmake-3.18-aarch64.tar.gz
-    tar -xvf cmake-3.18-aarch64.tar.gz
-    cp -ar cmake-3.18-aarch64 ${HOME}
-    PATH=${HOME}/cmake-3.18-aarch64/bin:$PATH
-    which cmake
-    cmake --version
 }
 
 maximize_ubuntu_github_actions_build_space() {