+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * RealSense sensor configuration, live capture and recording (with example and tutorial) (PR #2748)
 * Add mouselook for the legacy visualizer (PR #2551)
 * Add function to randomly downsample pointcloud (PR #3050)
+* Allow TriangleMesh with textures to be added (PR #3170)
+* Python property of open3d.visualization.rendering.Open3DScene `get_view` has been renamed to `view`.
 
 ## 0.11